--- client/src/Components/Contacts.js
@@ -1,18 +1,0 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-import './CSS/App.css';
-
-
-class Contacts extends Component {
-    render() {
-        return (
-            <>
-                <div className='item3'>
-                    <p>Esta é a página de contactos.</p>
-                </div>
-            </>
-        );
-    }
-}
-
-export default Contacts;
